// apps/web/src/api/__generated__/types/GetNotificationsResponseNotificationsNodesItemContent.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * sync
 * OpenAPI spec version: 0.0.1
 */
import type { GetNotificationsResponseNotificationsNodesItemContentActor } from './GetNotificationsResponseNotificationsNodesItemContentActor';

/**
 * Node Content
 */
export type GetNotificationsResponseNotificationsNodesItemContent = {
  /** Notification actor */
  actor?: GetNotificationsResponseNotificationsNodesItemContentActor;
  /** Creation timestamp */
  createdAt: string;
  /** Notification ID */
  id: number;
  /** Notification type */
  type: string;
  /** Notification status */
  status: string;
};